+++ src/Field/Radio.php
@@ -40,12 +40,31 @@ class Radio extends Checkbox {
     protected $choices = [];
 
     /**
+     * Whether an empty value should be allowed
+     *
+     * @var integer
+     */
+    protected $allow_null;
+
+
+    /**
      * Allow other choice
      *
      * @return self
      */
     public function allow_other_choice() {
         $this->other_choice = 1;
+
+        return $this;
+    }
+
+    /**
+     * Allow null value
+     *
+     * @return self
+     */
+    public function allow_null() {
+        $this->allow_null = 1;
 
         return $this;
     }